requirements.txt
accelerate
adapters
aiohttp
aiosignal
async-timeout
attrs
certifi
emoji
filelock
frozenlist
fsspec
huggingface-hub
charset-normalizer
idna
ipykernel
iterative-stratification
Jinja2
MarkupSafe
mpmath
multidict
networkx
numpy
packaging
pandas
polars==0.19.12 # 0.19.13 doesn't have pre-built wheels for linux
python-dateutil
pytz
pyyaml
torch==2.0.1 # cuda 12.1
torchmetrics
tqdm
transformers # FIXME doesn't work with adapters, probably?
trl
tweet-preprocessor
wandb
